Challenge 1 completed

+++ bamazonSchema.sql
@@ -7,7 +7,7 @@ CREATE TABLE products(
   item_id INT NOT NULL AUTO_INCREMENT,
   product_name VARCHAR(100) NOT NULL,
   department_name VARCHAR(50) NOT NULL,
-  price DOUBLE PRECISION default 0.00,
+  price DECIMAL (10,2) default 0.00,
   stock_quantity INT default 0,
   PRIMARY KEY (id)
 );
